✨ 회원탈퇴 컴포넌트 분리 (#118)
